cmd/libsnap: move apparmor-support to libsnap
The apparmor-support was supposed to be in libsnap-confine-private (long
time ago) but at the time I didn't move it there because unlike what the
name says, it was being used from tools other than snap-confine,
bringing along the libapparmor and libcap dependencies to places where
that was undesirable.

With shaving the yak in snap-discard-ns we can now do just that.
