+++ cmd/libsnap-confine-private/apparmor-support.h
